Move helper libs to libs folder
