+++ test/lib/helperFns.js
